Optimizations/Changed float numbers to 32bit

+++ src/Benchmarks/Image/RayTracingTask.hpp
@@ -1,23 +1,24 @@
 //
-//  Copyright (c) 2024  Ioannis Panagiotopoulos
+// Copyright (C) 2024-2025. Ioannis Panagiotopoulos
 //
-//  This program is free software: you can redistribute it and/or modify
-//  it under the terms of the GNU General Public License as published by
-//  the Free Software Foundation, either version 3 of the License, or
-//  (at your option) any later version.
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
 //
-//  This program is distributed in the hope that it will be useful,
-//  but WITHOUT ANY WARRANTY; without even the implied warranty of
-//  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-//  GNU General Public License for more details.
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
 //
-//  You should have received a copy of the GNU General Public License
-//  along with this program.  If not, see <https://www.gnu.org/licenses/>.
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 #ifndef ELPIDA_RAYTRACINGTASK_HPP
 #define ELPIDA_RAYTRACINGTASK_HPP
 
 #include "Elpida/Core/MicroTask.hpp"
+#include "Elpida/Core/Float.hpp"
 #include "Elpida/Graphics/FrameBuffer.hpp"
 #include "Elpida/Graphics/Scene.hpp"
 #include "Elpida/Graphics/Camera.hpp"
@@ -50,9 +51,9 @@ namespace Elpida
 		[[nodiscard]]
 		UniquePtr<Task> DoDuplicate() const override;
 	private:
-		Graphics::FrameBuffer<double> _frameBuffer;
-		Graphics::Scene<double> _scene;
-		Graphics::Camera<double> _camera;
+		Graphics::FrameBuffer<Float> _frameBuffer;
+		Graphics::Scene<Float> _scene;
+		Graphics::Camera<Float> _camera;
 		SharedPtr<AbstractTaskData> _inputData;
 		std::size_t _size;
 	};
