+++ src/Benchmarks/Math/MatrixInverse4Task.hpp
@@ -18,6 +18,7 @@
 #define ELPIDA_MATRIXINVERSE4TASK_HPP
 
 #include "Elpida/Core/MicroTask.hpp"
+#include "Elpida/Core/Float.hpp"
 #include "Elpida/Math/Matrix.hpp"
 
 namespace Elpida
@@ -48,8 +49,8 @@ namespace Elpida
 		[[nodiscard]]
 		UniquePtr<Task> DoDuplicate() const override;
 	private:
-		Math::Matrix<double, 4, 4> _matrix;
-		Math::Matrix<double, 4, 4> _input;
+		Math::Matrix<Float, 4, 4> _matrix;
+		Math::Matrix<Float, 4, 4> _input;
 		SharedPtr<AbstractTaskData> _inputData;
 	};
