+++ src/Benchmarks/StdLib/FloorTask.cpp
@@ -1,6 +1,23 @@
+//
+// Copyright (C) 2025. Ioannis Panagiotopoulos
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
 #include "FloorTask.hpp"
 
 #include "Elpida/Core/Repeat.hpp"
+#include "Elpida/Core/Float.hpp"
 
 namespace Elpida
 {
@@ -22,7 +39,7 @@ namespace Elpida
 
 	void FloorTask::DoRunImpl()
 	{
-		double y = 1561131321.1513123;
+		Float y = 1561131321.1513123;
 		Exec([&]()
 		{
 			REPEAT_10000(y = std::floor(y));
