+++ src/Benchmarks/StdLib/SqrtTask.cpp
@@ -1,7 +1,24 @@
+//
+// Copyright (C) 2025. Ioannis Panagiotopoulos
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
 #include "SqrtTask.hpp"
 
 #include <cmath>
 #include "Elpida/Core/Repeat.hpp"
+#include "Elpida/Core/Float.hpp"
 
 namespace Elpida
 {
@@ -22,7 +39,7 @@ namespace Elpida
 
 	void SqrtTask::DoRunImpl()
 	{
-		double x = 156113132146186435.1513123;
+		Float x = 156113132146186435.1513123;
 		Exec([&]()
 		{
 			REPEAT_10000(x = std::sqrt(x));
